feat : SoftDelete 구현

+++ src/main/kotlin/com/yoong/myissue/domain/comment/service/CommentService.kt
@@ -1,11 +1,13 @@
 package com.yoong.myissue.domain.comment.service
 
+import com.yoong.myissue.common.`class`.ValidAuthentication
 import com.yoong.myissue.domain.comment.dto.CreateCommentRequest
 import com.yoong.myissue.domain.comment.dto.UpdateCommentRequest
 import com.yoong.myissue.domain.comment.entity.Comment
 import com.yoong.myissue.domain.comment.repository.CommentRepository
 import com.yoong.myissue.domain.issue.service.ExternalIssueService
 import com.yoong.myissue.domain.member.service.ExternalMemberService
+import com.yoong.myissue.exception.`class`.BadRequestException
 import com.yoong.myissue.exception.`class`.ModelNotFoundException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -16,7 +18,8 @@ import org.springframework.transaction.annotation.Transactional
 class CommentService(
     private val commentRepository: CommentRepository,
     private val issueService: ExternalIssueService,
-    private val memberService: ExternalMemberService
+    private val memberService: ExternalMemberService,
+    private val validAuthentication: ValidAuthentication
 ){
 
     fun createComment(createCommentRequest: CreateCommentRequest, email: String): String {
@@ -33,15 +36,19 @@ class CommentService(
         return "댓글이 등록 되었습니다 이슈 번호 : ${createCommentRequest.issueId}, 코맨트 내용 : ${createCommentRequest.content}"
     }
 
-    fun updateComment(commentId: Long, updateCommentRequest: UpdateCommentRequest): String {
+    fun updateComment(commentId: Long, updateCommentRequest: UpdateCommentRequest, email: String): String {
+        val member = memberService.searchEmail(email)
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("댓글", commentId.toString())
+        if(!validAuthentication.isSame(member, comment)) throw BadRequestException("다른 사람의 댓글은 수정할 수 없습니다")
         comment.update(updateCommentRequest)
         commentRepository.save(comment)
         return "댓글 수정 완료 했습니다"
     }
 
-    fun deleteComment(commentId: Long): String {
+    fun deleteComment(commentId: Long, email: String): String {
+        val member = memberService.searchEmail(email)
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("댓글", commentId.toString())
+        if(!validAuthentication.isSame(member, comment)) throw BadRequestException("다른 사람의 댓글은 삭제할 수 없습니다")
         commentRepository.delete(comment)
         return "삭제가 완료 되었습니다 댓글 아이디 : $commentId"
     }
